# schemas/custom_gpt_action_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pd-asset-pipeline",
    "version": "1.0.0",
    "description": "FastAPI backend for turning public-domain archive pages into print-ready digital bundles."
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
    "/health": {
      "post": {
        "operationId": "health",
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {"type": "string"},
                    "service": {"type": "string"}
                  },
                  "required": ["status", "service"]
                }
              }
            }
          }
        }
      }
    },
    "/search_sources": {
      "post": {
        "operationId": "searchSources",
        "summary": "Search BHL and Internet Archive metadata",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchSourcesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Candidate source items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/extract_pages": {
      "post": {
        "operationId": "extractPages",
        "summary": "Extract pages from PDF or direct image URLs",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractPagesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Extracted image files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExtractedPage"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/build_bundle": {
      "post": {
        "operationId": "buildBundle",
        "summary": "Generate cleaned image set + print folders + zip",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildBundleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bundle metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bundle_id": {"type": "string"},
                    "local_path": {"type": "string"},
                    "download_url": {"type": "string"}
                  },
                  "required": ["bundle_id", "local_path", "download_url"]
                }
              }
            }
          }
        }
      }
    },
    "/bundle/{bundle_id}": {
      "get": {
        "operationId": "downloadBundle",
        "summary": "Download bundle ZIP",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "ZIP file"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SearchSourcesRequest": {
        "type": "object",
        "properties": {
          "query": {"type": "string"},
          "archive": {
            "type": "array",
            "items": {"type": "string", "enum": ["bhl", "internet_archive"]},
            "default": ["bhl", "internet_archive"]
          },
          "year_start": {"type": "integer"},
          "year_end": {"type": "integer"},
          "limit": {"type": "integer", "default": 10}
        },
        "required": ["query"]
      },
      "SearchResultItem": {
        "type": "object",
        "properties": {
          "title": {"type": "string"},
          "author": {"type": ["string", "null"]},
          "year": {"type": ["integer", "null"]},
          "source": {"type": "string"},
          "source_url": {"type": "string"},
          "item_id": {"type": "string"}
        },
        "required": ["title", "source", "source_url", "item_id"]
      },
      "ExtractPagesRequest": {
        "type": "object",
        "properties": {
          "pdf_url": {"type": "string"},
          "image_urls": {"type": "array", "items": {"type": "string"}},
          "page_numbers": {"type": "array", "items": {"type": "integer"}},
          "max_auto_pages": {"type": "integer", "default": 8}
        }
      },
      "ExtractedPage": {
        "type": "object",
        "properties": {
          "page_number": {"type": "integer"},
          "image_path": {"type": "string"},
          "thumbnail_path": {"type": "string"},
          "score": {"type": ["number", "null"]}
        },
        "required": ["page_number", "image_path", "thumbnail_path"]
      },
      "BuildBundleRequest": {
        "type": "object",
        "properties": {
          "bundle_name": {"type": "string"},
          "source_urls": {"type": "array", "items": {"type": "string"}},
          "extracted_page_paths": {"type": "array", "items": {"type": "string"}},
          "print_ratios": {
            "type": "array",
            "items": {"type": "string", "enum": ["2x3", "3x4", "4x5", "11x14"]},
            "default": ["2x3", "3x4", "4x5", "11x14"]
          },
          "cleanup_flags": {
            "type": "object",
            "additionalProperties": {"type": "boolean"}
          }
        },
        "required": ["bundle_name"]
      }
    }
  }
}